requirements.txt
numpy==1.22.4
Pillow==9.4.0
scikit_image==0.19.3
skimage==0.0
streamlit==1.12.0
streamlit_drawable_canvas==0.9.2
tensorflow==2.11.0
